Provide tuning on strictness of internal parser
Allow iter_paragraphs to be tuned as to whether it splits paragraphs where
the separator is a line containing whitespace. When working with
debian/control, iter_paragraphs must split on whitespace-only lines
(see #715558 for example). When working with Packages or Sources,
apt_pkg does not split on whitespace containing lines and so neither should
the internal parser.

(Closes: #913274)