Provide access to source, version in Packages
The Packages file may have an explicit Source field with the source package
name and optionally also the version specified in it; if these values are the
same as the binary package being described, they are omitted, meaning that
anything needing source package names and versions must have extra logic.

This patch provides uniform access to the source package name and source
package version for both the explicit and implicit case.

Closes: #934273