Provide dump() for Deb822ParagraphElements
This should enable some uses of `iter_paragraphs` to be "blindly"
convertable to the format preserving API.  Note that whitespace and
comments *between* paragraphs are lost in that case (if you do "per
paragraph" dumping rather than "file-level" dumping).  However, it can
be an acceptable loss for supporting both APIs at the same time with
minimal fuzz.

Signed-off-by: Niels Thykier <niels@thykier.net>