Rewrite OrderedSet as a LinkedHashSet
Compared to the original implementation, this rewrite *also* provides
O(1) for removal of field names.  Frankly, this rewrite only made
sense because we already had a suitable LinkedList implementation
available.

Additionally, it creates a good starting point for providing
specialized features such as "insert after" or "move to the end"
functionality which is used by `wrap-and-sort` and ebmutate`.

Signed-off-by: Niels Thykier <niels@thykier.net>