Have copyright.py pass type checks again
Moved the `RestrictedWrapper` into `debian.copyright` and made it
private to simplify the migration.  The class was not used outside
`python-debian` according to codesearch.d.o.  There are some external
users that fiddle with accessing `__data` directly but those should
still work unchanged after this.

This migration does *not* change the underlying mechanism for
accessing values as higher order types (like list of strings).
Accordingly, we still "nuke" all whitespace when someone updates the
`files` attribute of a `Files` paragraph.  Fixing that will come
later.

Signed-off-by: Niels Thykier <niels@thykier.net>