RTS parser: Speed up error detection
Instead of building the entire parse tree and then scanning it for
error tokens (or elements), we can just hook into the token stream
directly and abort on the first error.

This speeds up error detection even for valid files (as we save an
iteration over the entire AST in that case).

Signed-off-by: Niels Thykier <niels@thykier.net>