deb822: Make short-cut for internal-parser
During parsing, we would call `isinstance` roughly 3 times per line.
One of these cases happens via the `_AutoDecoder`'s `decode` method
called via the `_internal_parser` method.  However, in the
`_internal_parser` we know for sure that the input is bytes and with
this small bit of refactoring, we can skip the `isinstance` call.

Signed-off-by: Niels Thykier <niels@thykier.net>