Deb822.validate_input: Reorder two conditions
If there are no `\n` characters in the value, then we know the field
value does not end on one as well.  This has the advantage that all
single line fields (such as `Package: foo`) are now checked with a
single `if` rather than going through two.

A lot of fields turn out to be single line ones. Especially in
`Packages` files, where "redundant" whitespace in Depends (etc.) are
stripped.

Signed-off-by: Niels Thykier <niels@thykier.net>