RTS parser: Tighten some types to be read-only
Nothing was doing mutations via these attributes and doing mutations
via these attributes were not recommeded (or supported), so this is
mostly about correctness.

However, in the next commit, a "size" cache is added to elements and
forcing these attributes to be immutable makes it easier to reason why
there is no cache reset for these objects.

Signed-off-by: Niels Thykier <niels@thykier.net>