RTS Locable: Improve feature set a bit to match usage better
Common use-case of the locatable is to define the range of the
locatable in the file via some cached reference point. Previously,
this took a lot of effort. Something like:

```python
    parent_pos_in_file: TEPosition = ...
    child: Locatable = ...
    pos_in_parent = child.position_in_parent()
    pos_in_file = pos_in_parent.relative_to(parent_pos_in_file)
    range_in_file = TERange.from_position_and_size(
      pos_in_file,
      child.te_size()
    )
```

With this commit, we can shorten that a bit to:

```python
    parent_pos_in_file: TEPosition = ...
    child: Locatable = ...
    range_in_parent = child.range_in_parent()
    range_in_file = range_in_parent.relative_to(parent_pos_in_file)
```

Which takes out an entire intermediate step.  Also, if you need the
position with leading comments, you only have to provide it once
(`range_in_parent`) rather than twice (`position_in_parent` +
`te_size`). This leaves less room for errors as you would no longer
get them out of sync.

To avoid confusion with the new `TERange.relative_to`
vs. `TERange.rebase`, the `rebase` method has been renamed to
`as_size` and it no longer accepts a base point.  This simplifies its
feature to "always reduce the range into a size", which makes it
simplier to explain.

Signed-off-by: Niels Thykier <niels@thykier.net>