Locatable: Drop `skip_leading_comment` kwarg
It leads to many issues with getting the correct line number for
fields that have comments.

At some point, we should probably look into easy supporting iterating
over all KVPairs in a stanza along with their ranges to avoid the
`O(n²)` problem that happens with the current API if you want to get
the range for each field in the stanza.